feat: move map display component to design-system
- details page URL is now assembled outside of the component
- no longer uses i18n
