+++ package.json
@@ -25,7 +25,6 @@
     "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "13.4.0",
     "@testing-library/user-event": "13.5.0",
-    "@types/leaflet.markercluster": "^1.5.6",
     "@types/node": "22.18.10",
     "@types/react": "18.0.27",
     "@types/react-dom": "18.0.10",
@@ -39,8 +38,6 @@
     "i18next": "25.6.0",
     "i18next-http-backend": "3.0.2",
     "js-md5": "0.8.3",
-    "leaflet": "^1.9.4",
-    "leaflet.markercluster": "^1.5.3",
     "moment-timezone": "0.5.43",
     "react-bootstrap": "2.7.2",
     "react-bootstrap-icons": "1.11.4",
@@ -135,7 +132,6 @@
     "@storybook/test-runner": "0.23.0",
     "@testing-library/cypress": "10.1.0",
     "@types/chai-as-promised": "7.1.5",
-    "@types/leaflet": "^1.9.21",
     "@types/node-sass": "4.11.3",
     "@types/sinon": "10.0.13",
     "@typescript-eslint/eslint-plugin": "5.51.0",
