+++ packages/design-system/package.json
@@ -33,6 +33,11 @@
     "test:coverage": "nyc check-coverage"
   },
   "dependencies": {
+    "@types/leaflet.markercluster": "^1.5.6",
+    "leaflet": "^1.9.4",
+    "leaflet.markercluster": "^1.5.3",
+    "react-leaflet": "^4.2.1",
+    "react-loading-skeleton": "3.3.1",
     "@dnd-kit/core": "6.1.0",
     "@dnd-kit/modifiers": "^7.0.0",
     "@dnd-kit/sortable": "8.0.0",
@@ -53,6 +58,7 @@
     "vite-plugin-istanbul": "4.0.1"
   },
   "devDependencies": {
+    "@types/leaflet": "^1.9.21",
     "@storybook/react": "8.6.14",
     "@storybook/test-runner": "0.23.0",
     "@testing-library/cypress": "10.1.0",
