+++ packages/design-system/src/lib/components/map/DatasetsMap.module.scss
@@ -1,4 +1,4 @@
-@import 'node_modules/@iqss/dataverse-design-system/src/lib/assets/styles/design-tokens/typography.module';
+@import '../../assets/styles/design-tokens/typography.module';
 
 .map-wrapper {
   border: 1px solid #dee2e6;
